backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
langchain==0.3.26
langchain-google-genai==2.0.10
langchain-community==0.3.27
google-generativeai==0.8.5
pandas==2.3.1
numpy==2.3.1
python-dotenv==1.1.1
faiss-cpu>=1.7.4 